+++ tests/Unit/Test.php
@@ -18,6 +18,7 @@ use Balloon\Server;
 use Helmich\MongoMock\MockDatabase;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
+use Psr\SimpleCache\CacheInterface;
 
 abstract class Test extends TestCase
 {
@@ -50,6 +51,7 @@ abstract class Test extends TestCase
             self::getMockDatabase(),
             $this->createMock(NodeFactory::class),
             $this->createMock(LoggerInterface::class),
+            $this->createMock(CacheInterface::class),
             $this->createMock(Hook::class),
             $acl
         );