+++ CHANGELOG.md
@@ -5,11 +5,12 @@
 
 * API: [FIX] memory leak in delta request specifc node recursive #233
 * CORE: [CHANGE] Add new field pointer which gets used to recursively resolve child nodes #234
+* CORE: [FIX] Share reference may be created twice #231
 
 
 ## 2.2.1
 **Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
-**Date**: Tue Dez 18 11:35:32 CET 2018\
+**Date**: Tu  Dez 18 11:35:32 CET 2018\
 
 * API: [FIX] Initial delta may contain nodes with the same path with different id (double share references) #230
 * API: [FIX] Delta with specific node id may result in empty responses but has_more:true #229